refactor(Core/sWorld): improve singleton management
